Improve perceived performance of front page
Since all available elements are shown on the front page there is a
substantial delay before the user sees a response from the server.
Switching the return type of the data returned by the load function to a
Promise enables streaming and defers rendering of the elements to the
client.