Re-enable counting of all foci in the result set
This allows us to show the count in the assignee filter. As a downside
foci stemming from objects which are not displayed might show up in the
filter bar. Hence we need to apply a filter to the objects being
counted to only count foci from objects intended for display.