Ensure drag handle is only displayed if table is wide enough
We observed that sorting the table rows is only possible when more than
half of a row's width is visible. Apparently this has to do with the
approach taken by svelted-dnd-action to use the center of the dragged
object to determine intersections. Configuring the action with
centreDraggedOnCursor does not help in this case, neither does changing
the width of the dragged row with transformDraggedElement.